chore(refactor): move CollectionOverwrite to src/overwrite/types.ts (refs #47)
