+++ src/overwrite/overwrite-store.ts
@@ -1,7 +1,7 @@
 import { existsSync, readFileSync } from 'fs';
 import { join } from 'path';
 
-import type { CollectionOverwrite } from '../types';
+import type { CollectionOverwrite } from './types';
 import { getDataDir } from '../config/data-dir';
 import { parseOverwrite } from './overwrite';
