+++ test/unit/helpers/merge.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import type { CollectionOverwrite } from '@/types'
+import type { CollectionOverwrite } from '@/overwrite/types'
 import type { SourceCollection } from '@/source/types'
 import { merge } from '@/enrichment/merge'
